+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Logging Module
 - Base Networking Module
 - UIColor from HEX
+- Keyboard Layout Guide
 
 ### Changed